// examples/cstabletops/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  "admin/audit": typeof admin_audit;
  "admin/authorization": typeof admin_authorization;
  auth: typeof auth;
  authorization: typeof authorization;
  http: typeof http;
  scaffold: typeof scaffold;
  tasquencer: typeof tasquencer;
  "workflows/cstabletops/api": typeof workflows_cstabletops_api;
  "workflows/cstabletops/authSetup": typeof workflows_cstabletops_authSetup;
  "workflows/cstabletops/db": typeof workflows_cstabletops_db;
  "workflows/cstabletops/definition": typeof workflows_cstabletops_definition;
  "workflows/cstabletops/exercises": typeof workflows_cstabletops_exercises;
  "workflows/cstabletops/helpers": typeof workflows_cstabletops_helpers;
  "workflows/cstabletops/migrations": typeof workflows_cstabletops_migrations;
  "workflows/cstabletops/scopes": typeof workflows_cstabletops_scopes;
  "workflows/cstabletops/workItems/authHelpers": typeof workflows_cstabletops_workItems_authHelpers;
  "workflows/cstabletops/workItems/cardTasks": typeof workflows_cstabletops_workItems_cardTasks;
  "workflows/metadata": typeof workflows_metadata;
}>;

/**
 * A utility for referencing Convex functions in your app's public API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;